Revert grayscale to floating changes

+++ IOSAccessAssessment/Segmentation/GrayscaleToColorCIFilter.swift
@@ -12,7 +12,7 @@ import MetalKit
 
 class GrayscaleToColorCIFilter: CIFilter {
     var inputImage: CIImage?
-    var grayscaleValues: [UInt8] = []
+    var grayscaleValues: [Float] = []
     var colorValues: [CIColor] = []
     
     // Metal-related properties
@@ -73,7 +73,7 @@ class GrayscaleToColorCIFilter: CIFilter {
             return nil
         }
         
-        let grayscaleBuffer = self.device.makeBuffer(bytes: grayscaleValues, length: grayscaleValues.count * MemoryLayout<UInt8>.size, options: [])
+        let grayscaleBuffer = self.device.makeBuffer(bytes: grayscaleValues, length: grayscaleValues.count * MemoryLayout<Float>.size, options: [])
         let colorBuffer = self.device.makeBuffer(bytes: colorValues.map { SIMD3<Float>(Float($0.red), Float($0.green), Float($0.blue)) }, length: colorValues.count * MemoryLayout<SIMD3<Float>>.size, options: [])
         
         commandEncoder.setComputePipelineState(self.pipeline)
